+++ src/de/deserializer.rs
@@ -78,12 +78,25 @@ impl core::fmt::Display for Error {
 pub struct MqttDeserializer<'a> {
     buf: &'a [u8],
     index: usize,
+    next_pending_length: Option<usize>,
 }
 
 impl<'a> MqttDeserializer<'a> {
     /// Construct a deserializer from a provided data buffer.
     pub fn new(buf: &'a [u8]) -> Self {
-        Self { buf, index: 0 }
+        Self {
+            buf,
+            index: 0,
+            next_pending_length: None,
+        }
+    }
+
+    /// Override the next binary bytes read with some pre-determined size.
+    ///
+    /// # Args
+    /// * `len` - The known length of the next binary data blob.
+    pub fn set_next_pending_length(&mut self, len: usize) {
+        self.next_pending_length.replace(len);
     }
 
     /// Attempt to take N bytes from the buffer.
@@ -121,6 +134,11 @@ impl<'a> MqttDeserializer<'a> {
     /// Read a variable-length integer from the data buffer.
     pub fn read_varint(&mut self) -> Result<u32, Error> {
         self.read_u32_varint()
+    }
+
+    /// Determine the number of bytes that were deserialized.
+    pub fn deserialized_bytes(&self) -> usize {
+        self.index
     }
 
     /// Extract any remaining data from the buffer.
@@ -188,8 +206,11 @@ impl<'de, 'a> serde::de::Deserializer<'de> for &'a mut MqttDeserializer<'de> {
     }
 
     fn deserialize_bytes<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value, Self::Error> {
-        let length = self.read_u16()?;
-        let bytes: &'de [u8] = self.try_take_n(length as usize)?;
+        let length = match self.next_pending_length.take() {
+            Some(length) => length,
+            None => self.read_u16()? as usize,
+        };
+        let bytes: &'de [u8] = self.try_take_n(length)?;
         visitor.visit_borrowed_bytes(bytes)
     }
 
@@ -207,12 +228,17 @@ impl<'de, 'a> serde::de::Deserializer<'de> for &'a mut MqttDeserializer<'de> {
     }
 
     fn deserialize_seq<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value, Self::Error> {
-        // Sequences are always prefixed with the number of bytes contained within them encoded as
-        // a variable-length integer.
-        let len = self.read_varint()?;
+        // Sequences, which are properties, are always prefixed with the number of bytes contained
+        // within them encoded as a variable-length integer.
+        let length = self.read_varint()? as usize;
+
+        // If the properties are read as a binary blob, we already know the size and we don't
+        // want to read a u16-prefixed size.
+        self.set_next_pending_length(length);
+
         visitor.visit_seq(SeqAccess {
             deserializer: self,
-            len: len as usize,
+            length,
         })
     }
 
@@ -345,7 +371,7 @@ impl<'a, 'de: 'a> serde::de::SeqAccess<'de> for ElementAccess<'a, 'de> {
 /// Structure used to access a specified number of bytes.
 struct SeqAccess<'a, 'de: 'a> {
     deserializer: &'a mut MqttDeserializer<'de>,
-    len: usize,
+    length: usize,
 }
 
 impl<'a, 'de: 'a> serde::de::SeqAccess<'de> for SeqAccess<'a, 'de> {
@@ -355,15 +381,18 @@ impl<'a, 'de: 'a> serde::de::SeqAccess<'de> for SeqAccess<'a, 'de> {
         &mut self,
         seed: V,
     ) -> Result<Option<V::Value>, Error> {
-        if self.len > 0 {
+        if self.length > 0 {
             // We are deserializing a specified number of bytes in this case, so we need to track
             // how many bytes each serialization request uses.
             let original_remaining = self.deserializer.len();
             let data = DeserializeSeed::deserialize(seed, &mut *self.deserializer)?;
-            self.len = self
-                .len
+            self.length = self
+                .length
                 .checked_sub(original_remaining - self.deserializer.len())
                 .ok_or(Error::InsufficientData)?;
+
+            // Since some data was just read, we no longer know the size of the binary data block.
+            self.deserializer.next_pending_length.take();
 
             Ok(Some(data))
         } else {